change ID type

+++ models/models.go
@@ -1,12 +1,12 @@
 package models
 
 type Customer struct {
-	ID        int    `json:"id"`
-	Name      string `json:"name"`
-	Role      string `json:"role"`
-	Email     string `json:"emil"`
-	Phone     string `json:"phone"`
-	Contacted bool   `json:"contacted"`
+	ID        float64 `json:"id"`
+	Name      string  `json:"name"`
+	Role      string  `json:"role"`
+	Email     string  `json:"emil"`
+	Phone     string  `json:"phone"`
+	Contacted bool    `json:"contacted"`
 }
 
 // type Customers struct {
